backend/requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
qrcode==8.2
sqlparse==0.5.5
tzdata==2025.3
